spec: auto-generate decode/encode mapping
This uses a higher-level description inside the build system in
order to generate most of the tedious mapping code from a more succinct
description. Untested other than compilation!

Currently it can generate the entire RV32I Base Instruction Set, except
for the FENCE instruction.

This also removes the execute/assembly mapping clauses for now. Cleaning
those up and returning them (and whether or not execute() will be
generated) can be figured out in the future.

Signed-off-by: Austin Seipp <aseipp@pobox.com>