build: shim bake.hs, move all logic into src/mk
Previously, the buildExe target in release.nix just blindly used ./. as
the source path to keep thing simple -- because the logic existed in
src/mk as well as bake.hs.

This had a negative impact on using the compiled build system for nix
builds: because any source change anywhere would result in a build
system recompile, even if the Shake code wasn't changed!

This moves bake.hs into src/mk/Bake.hs, and ./bake.hs is now simply
a wrapper script that calls Bake.main. Furthermore, the release.nix
buildExe target has been modified to track ./src/mk as the build
directory. With that, we can actually still compile everything easily.

This does make it possible for a small build inconsistency window to
occur, e.g. where only bake.hs is modified and a rebuild is run. But
that would require logic changes to bake.hs, and avoiding those is the
whole point!

Signed-off-by: Austin Seipp <aseipp@pobox.com>