libfirm: fix _start.S.o section ordering bug
Previously the linker script for libfirm just specified that the
'start*' object should occur first in the resulting linked output --
this is necessary so the reset vector lands on address 0x00000000.
However, this only seemed to work coincidentally: when using more
objects, the resulting ELF file had misordered entry code.

The basic problem _seems_ to be that:

  - LD expects the path of an object listed in a script to *exactly*
    match the filename passed on the command line, modulo globbing,
    but including any relative paths included on the command line.

  - If an object matching some glob isn't found, i.e. the glob returns
    an empty set of paths, no warning is emitted for this.

These restrictions make sense, more or less. But the first issue arises
because we use a "non recursive" build that always executes from the
project top directory -- that is we always give a full relative path to
the build objects, or move to other directories. This means we do not
invoke GCC like:

    gcc -o foo.elf _start.S.o

but more like:

    gcc -o ... build/src/libfirm/_start.S.o

so the path to _start.S.o is relative, not in the CWD. This means globs
like 'start*' simply do not match because they do not include the
relative prefix.

The fix is luckily very simple: just move the glob to be a prefix,
rather than a suffix. Now, we always order linker output sections such
that any object named '*_start.S.o' appears first.

Signed-off-by: Austin Seipp <aseipp@pobox.com>