nix: build a rv32i GCC compiler, not rv32imc
Previously the GCC cross compiler was configured as an 'rv32imc'
compiler -- that is, it was allowed free use of both hard-mult
instructions and, more importantly, compressed instructions.

For certain bare metal programs, this isn't a problem -- resulting
object files can be made completely 'standalone' and using the
-march=rv32i option, you can control *individual* compiler invocations
and restrict the use of e.g. compressed instructions in the output. If
this works fine, then you do not need anything else.

But, if programs like firmware need functions from, say, libgcc --
because they use operations like div/mod, or something that could not be
emitted "inline" by the code generator -- then not all is well. Because
libgcc is created at GCC-build-time using the 'rv32imc' assumption, it
will be able to freely emit compressed instructions in the resulting
output files. The end result is that if you mix rv32imc libgcc with pure
rv32i objects, you will still get compressed instructions at the end.

Naturally, this is problematic since we don't decode/support
decompressed instructions yet. Luckily, the fix is trivial!

Signed-off-by: Austin Seipp <aseipp@pobox.com>