test: rework minimal test firmware
The previous test firmware (stolen from picorv32) was mostly cobbled
together to get an idea of what the build system needed to support it
and build ELF images. Since the simulator is now (barely) working
(hobbling along), this reworks the firmware to get rid of most of the
PicoRV32 assumptions (e.g. IRQ lines) and gets a very minimal 'Sieve'
example building. Notably:

  - Kill all the old Bake cruft that supported the 'old' build rules.
  - Build the firmware using the new Bake.CC module rules
  - Kill all the old unrelated crap -- linker scripts, custom IRQ ops,
    etc
  - Get a simple sieve example and 'stats' example linked in (though the
    stats are obviously garbage, since rdcycle isn't supported yet!)

Notably, this *does not* yet include, link, or run the actual RISC-V
opcode tests (written in assembly), but those should be pretty easy to
add on top of this (I hope).

Signed-off-by: Austin Seipp <aseipp@pobox.com>