t: hook up asm tests into smoke.elf firmware
This re-enables compilation of the RISC-V assembly language tests, and
also sets up the necessary infrastructure to link them into the
smoke.elf object which should perform said smoke tests.

Currently all the tests are disabled, so the Sieve and stats functions
are run, but that's all.

This also rewrites the test firmware main function into a small
assembler snippet (because the test harness requires us to generate _ret
symbols for it.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>