bake: rename C/caml backend outputs
The primary development platform is the C emulator, so putting it under
`.opt` as a filename suffix is pointless -- it's just extra characters
to type. The static emulator targets output binaries named ~cruise~ as
well based on the C source code. And we don't even have build rules to
create images for the OCaml backend!

Ultimately that's what people will use, so we might as well just make it
easier to type and refer to.

The Caml backend is now under the name build/cruise.ref -- this required
working around a bug in the Sail Caml backend, which doesn't handle
filename characters like '.' properly.

Signed-off-by: Austin Seipp <aseipp@pobox.com>