nix: add useful cache-deps derivation and script
This derivation contains a list of paths which has all the build
dependencies, including the dependencies for the static emulator
compiler. There's an accompanying script to get everything on stdout all
at once.

The intention here is to just have a list of paths you can push around,
e.g. with this, you can upload all build artifacts and build
dependencies to Cachix via:

    $ ./src/etc/cache-deps.sh | cachix push <CACHE>

Signed-off-by: Austin Seipp <aseipp@pobox.com>