github: retry nix-build docker interaction
GitHub Actions lets you keep things around in the working directory.
Ideally, we'd like to run a Nix build, then take the results and keep
them for future actions. Unfortunately, when the result is a symlink,
this breaks stuff: every action runs in a *separate* container, so by
the time the nix container is destroyed the symlink is dead and invalid.

Instead, just always build whatever is asked, but DO NOT keep the
results (e.g. any result* symlinks) -- this will just make sure whatever
build you specify passes.

As a special case, if a Docker target is specified, then a 3rd argument
to the action allows you to specify an output file: this output file
will be the target of a "cp" operation that copies the Docker tarball
out of the store. This should allow actual 'docker load' commands to
work.

With that in place, we enable shell linting for the new entrypoint.sh
script, and try loading the Docker images into a new action. This should
allow us to also do image publishing, if it works.

Signed-off-by: Austin Seipp <aseipp@pobox.com>