github: add empty line to restart CI [NFC]
The CI process failed due to an external issue with the Docker login
Action, but this was a bug in the action itself, not our usage of it, I
think.

Add an empty line to trigger a new build process, to see if a workaround
can fix it (apparently the Beta version of GH Actions will not let you
re-run a workflow).

Signed-off-by: Austin Seipp <aseipp@pobox.com>