github: sync wasm-html build to Backblaze B2
With this, every commit to master (for now) will built and be uploaded.
In the future, this URL will probably change (we'll want to add proper
MIME types for .wasm files) and we may want to sync different versions
to different directories for preservation/versioning/testing purposes.

Right now, we use rclone and just rely on it to do 'soft deletes' to a
remote bucket, where old versions of the file are still available, not
really deleted. Because the WASM builds are relatively small (~1.5mb),
it's not very costly to do this. Besides -- I already have ~100s of GBs
of data in B2, so the cost of this is dwarfed anyway. (Even assuming
thousands of commits/syncs, it would be a while before even reaching the
10GB free limit for other accounts.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>